+++ spx-gui/src/components/ui/index.ts
@@ -1,6 +1,6 @@
 export { default as UIConfigProvider, useUIVariables } from './UIConfigProvider.vue'
 export { default as UICard } from './UICard.vue'
-export { default as UIPanelHeader } from './UIPanelHeader.vue'
+export { default as UICardHeader } from './UICardHeader.vue'
 export { default as UIButton } from './UIButton.vue'
 export { default as UIDropdown } from './UIDropdown.vue'
 export { UIMenu, UIMenuGroup, UIMenuItem } from './menu'
